requirements.txt
matplotlib==3.7.1
MetaTrader5==5.0.44
pandas==1.5.3
python-dotenv==1.0.0
pytz==2022.7.1
scipy==1.10.1
